requirements.txt
boto3==1.20.9; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
botocore==1.23.9; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
brotli==1.0.9
click==8.0.3; python_version >= "3.6"
colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.6" and python_full_version >= "3.5.0"
contextlib2==21.6.0; python_version >= "3.6"
crispy-forms-foundation==0.8.0
dj-database-url==0.5.0
django-basic-auth-ip-whitelist==0.2.1; python_version >= "3.4"
django-crispy-forms==1.8.1
django-csp==3.5
django-extensions==2.1.3
django-referrer-policy==1.0
django-storages==1.7.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
django==2.1.7; python_version >= "3.5"
entrypoints==0.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "2.7"
factory-boy==2.11.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
faker==1.0.5; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
flake8==3.7.7; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
humanize==3.12.0; python_version >= "3.6"
isort==4.3.15; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
jmespath==0.10.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
mccabe==0.6.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
psutil==5.8.0; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
psycopg2-binary==2.8.6; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
pycodestyle==2.5.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
pyflakes==2.1.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
python-dateutil==2.8.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pytz==2021.3; python_version >= "3.5"
pyyaml==6.0; python_version >= "3.6"
s3transfer==0.5.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
schema==0.7.4
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
sqlalchemy-utils==0.33.1
sqlalchemy==1.2.5
stellar==0.4.5
text-unidecode==1.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
urllib3==1.26.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.6"
werkzeug==0.14.1
whitenoise==4.1.2
